DRYed up our controllers
