+++ app/controllers/movies_controller.rb
@@ -1,5 +1,4 @@
 class MoviesController < ApplicationController
-
   def index
     movies = Movie.all
 
@@ -28,18 +27,16 @@ class MoviesController < ApplicationController
     end
   end
 
-  def copies_out #list of copies checked out
-    movie = Movie.find_by(id: params[:id])
+  private
 
-    #first check if the movie is even rented
-    if movie.copies_out?
-      rentals = movie.rentals
+    def movie_params
+      params.permit(:title, :overview, :release_date, :inventory, :available_inventory, :sort, :n, :p)
+    end
 
-      rentals = rentals.select do |rental|
-        rental.checked_out?
-      end
+    def rentals_list(criteria)
+      movie = Movie.find_by(id: params[:id].to_i)
 
-      list = rentals.map do |rental|
+      list = movie.send(criteria).map do |rental|
         {
           customer_id: rental.customer_id,
           checkout_date: rental.checkout_date,
@@ -52,34 +49,5 @@ class MoviesController < ApplicationController
       list = sort_and_paginate(list, movie_params)
 
       render json: list.as_json, status: :ok
-    else
-      render json: { message: "There are no copies currently out" }
-    end
-  end
-
-  def history
-    movie = Movie.find_by(id: params[:id])
-    rentals = movie.rentals
-
-    list = rentals.map do |rental|
-      if !rental.checked_out
-        { customer_id: rental.customer_id,
-          checkout_date: rental.checkout_date,
-          due_date: rental.due_date,
-          name: rental.customer.name,
-          postal_code: rental.customer.postal_code
-        }
-      end
-    end
-
-    list = sort_and_paginate(list, movie_params)
-
-    render json: list.as_json, status: :ok
-  end
-
-  private
-
-    def movie_params
-      params.permit(:title, :overview, :release_date, :inventory, :available_inventory, :sort, :n, :p)
     end
 end
