+++ app/controllers/rentals_controller.rb
@@ -1,7 +1,4 @@
-require 'date'
-
 class RentalsController < ApplicationController
-
   def checkout
     rental = Rental.new(rental_params)
 
@@ -17,6 +14,7 @@ class RentalsController < ApplicationController
       rental.movie_id == params[:movie_id].to_i && rental.customer_id == params[:customer_id].to_i
     end
 
+    # checkin the oldest rental
     if rentals.length >= 1
       rental = rentals.sort_by { |rental| rental.due_date }.first
     else
@@ -40,13 +38,15 @@ class RentalsController < ApplicationController
     end
 
     list = overdues.map! do |overdue|
-      { movie_id: overdue.movie_id,
-      customer_id: overdue.customer_id,
-      checkout_date: overdue.checkout_date,
-      due_date: overdue.due_date,
-      title: overdue.movie.title,
-      name: overdue.customer.name,
-      postal_code: overdue.customer.postal_code }
+      {
+        movie_id: overdue.movie_id,
+        customer_id: overdue.customer_id,
+        checkout_date: overdue.checkout_date,
+        due_date: overdue.due_date,
+        title: overdue.movie.title,
+        name: overdue.customer.name,
+        postal_code: overdue.customer.postal_code
+      }
     end
 
     list = sort_and_paginate(list, rental_params)
