+++ app/models/customer.rb
@@ -24,15 +24,15 @@ class Customer < ApplicationRecord
     end
   end
 
-  def current_rentals
-    return self.rentals.select do |rental|
-      rental.checked_out
-    end
-  end
-
-  def historical_rentals
-    return self.rentals.select do |rental|
-      !rental.checked_out
-    end
-  end
+  # def current_rentals
+  #   return self.rentals.select do |rental|
+  #     rental.checked_out
+  #   end
+  # end
+  #
+  # def historical_rentals
+  #   return self.rentals.select do |rental|
+  #     !rental.checked_out
+  #   end
+  # end
 end
