+++ config/routes.rb
@@ -8,7 +8,7 @@ Rails.application.routes.draw do
 
   get "/movies", to: "movies#index", as: "movies"
   get "/movies/:id", to: "movies#show", as: "movie"
-  get "/movies/:id/current", to: "movies#copies_out", as: "movies_out"
+  get "/movies/:id/current", to: "movies#current", as: "movies_out"
   get "/movies/:id/history", to: "movies#history", as: "rental_history"
   post "/movies", to: "movies#create"
 
@@ -16,82 +16,3 @@ Rails.application.routes.draw do
   post "/rentals/check-out", to: "rentals#checkout", as: "rental_checkout"
   post "/rentals/check-in", to: "rentals#checkin", as: "rental_checkin"
 end
-
-
-=begin
-
-`GET /customers`
-List all customers
-
-Fields to return:
-- `id`
-- `name`
-- `registered_at`
-- `postal_code`
-- `phone`
-- `movies_checked_out_count`
- - This will be 0 unless you've completed optional requirements
-
-#### `GET /movies`
-List all movies
-
-Fields to return:
-- `id`
-- `title`
-- `release_date`
-
-#### `GET /movies/:id`
-Look a movie up by `id`
-
-URI parameters:
-- `id`: Movie identifier
-
-Fields to return:
-- `title`
-- `overview`
-- `release_date`
-- `inventory` (total)
-- `available_inventory` (not currently checked-out to a customer)
- - This will be the same as `inventory` unless you've completed the optional endpoints.
-
-#### `POST /movies`
-Create a new movie in the video store inventory.
-
-Upon success, this request should return the `id` of the movie created.
-
-Request body:
-
-| Field         | Datatype            | Description
-|---------------|---------------------|------------
-| `title` | string             | Title of the movie
-| `overview` | string | Descriptive summary of the movie
-| `release_date` | string `YYYY-MM-DD` | Date the movie was released
-| `inventory` | integer | Quantity available in the video store
-
-### Wave 3: Rentals
-
-Wave 2 focused on working with customers and movies. With these endpoints you can extend the functionality of your API to allow managing the rental process.
-
-#### `POST /rentals/check-out`
-Check out one of the movie's inventory to the customer. The rental's check-out date should be set to today, and the due date should be set to a week from today.
-
-**Note:** Some of the fields from wave 2 should now have interesting values. Good thing you wrote tests for them, right... right?
-
-Request body:
-
-| Field         | Datatype            | Description
-|---------------|---------------------|------------
-| `customer_id` | integer             | ID of the customer checking out this film
-| `movie_id`    | integer | ID of the movie to be checked out
-
-#### `POST /rentals/check-in`
-Check in one of a customer's rentals
-
-Request body:
-
-| Field         | Datatype | Description
-|---------------|----------|------------
-| `customer_id` | integer  | ID of the customer checking in this film
-| `movie_id`    | integer | ID of the movie to be checked in
-
-=end
